refactor(be): 주문 레포지터리 커서 기반 조회 로직 개선
* 코드 중복 제거
* 가독성 개선
* QueryDSL limit() 인자 타임 명세에 맞게 관련 코드 수정

+++ backend/src/main/java/com/deliveranything/domain/order/controller/CustomerOrderController.java
@@ -52,7 +52,7 @@ public class CustomerOrderController {
   public ResponseEntity<ApiResponse<CursorPageResponse<OrderResponse>>> getAll(
       @AuthenticationPrincipal SecurityUser securityUser,
       @RequestParam(required = false) String nextPageToken,
-      @RequestParam(defaultValue = "10") int size
+      @RequestParam(defaultValue = "10") long size
   ) {
     return ResponseEntity.ok().body(ApiResponse.success("소비자 전체 주문 내역 조회 성공",
         customerOrderService.getCustomerOrdersByCursor(securityUser.getCurrentActiveProfileIdSafe(),
@@ -84,7 +84,7 @@ public class CustomerOrderController {
   public ResponseEntity<ApiResponse<CursorPageResponse<OrderResponse>>> getCompletedOrders(
       @AuthenticationPrincipal SecurityUser securityUser,
       @RequestParam(required = false) String nextPageToken,
-      @RequestParam(defaultValue = "10") int size
+      @RequestParam(defaultValue = "10") long size
   ) {
     return ResponseEntity.ok().body(ApiResponse.success("배달 완료된 소비자 주문 조회 성공",
         customerOrderService.getCompletedOrdersByCursor(
